# Move resource version from metric label to metric number value