These metrics expose the resource version as a string in the label. This value can change often and would therefore create huge cardinality. This PR changes it to be the number value of the metric itself instead.

I checked with API team and while formally it is not guaranteed that this is always a numerical value, majority of the ecosystem assumes this is the case.

cc @tariq1890 @brancz 

Fixes https://github.com/kubernetes/kube-state-metrics/issues/982