@LiliC 2019-12-11
Benchmark results:
https://travis-ci.org/kubernetes/kube-state-metrics/jobs/623616433#L541-L566

```
### Result
old=master new=HEAD
benchmark                                       old ns/op       new ns/op       delta
BenchmarkKubeStateMetrics/GenerateMetrics-2     1001245093      1000977482      -0.03%
BenchmarkKubeStateMetrics/MakeRequests-2        19276868580     17907123355     -7.11%
BenchmarkPodStore-2                             52782           50941           -3.49%
BenchmarkMetricWrite/value-1-2                  919             880             -4.24%
BenchmarkMetricWrite/value-35.7-2               1080            1052            -2.59%

benchmark                                    old MB/s     new MB/s     speedup
BenchmarkKubeStateMetrics/MakeRequests-2     538.59       578.61       1.07x
benchmark                                       old allocs     new allocs     delta
BenchmarkKubeStateMetrics/GenerateMetrics-2     849192         844208         -0.59%
BenchmarkKubeStateMetrics/MakeRequests-2        453942         454000         +0.01%
BenchmarkPodStore-2                             523            523            +0.00%
BenchmarkMetricWrite/value-1-2                  7              7              +0.00%
BenchmarkMetricWrite/value-35.7-2               7              7              +0.00%

benchmark                                       old bytes       new bytes       delta
BenchmarkKubeStateMetrics/GenerateMetrics-2     87317896        86970760        -0.40%
BenchmarkKubeStateMetrics/MakeRequests-2        26519171184     26518128840     -0.00%
BenchmarkPodStore-2                             25984           25984           +0.00%
BenchmarkMetricWrite/value-1-2                  536             536             +0.00%
BenchmarkMetricWrite/value-35.7-2               536             536             +0.00%
```