@brancz 2019-12-11
Very nice!

/lgtm
/hold

Putting on hold and I'll let @tariq1890 decide whether to include this as a next release candidate or whether we want to create a follow up 1.9.1 release.