@brancz 2019-12-11
Right the docs do need updating. It's a little weird to change them to experimental now, but realistically we're dependent on kubernetes/etcd behavior here, so if that would break this metric would break, so I'm not sure STABLE was ever the right thing/possible.