@tariq1890 2019-12-12
My preference is to have this in the rc itself. Let’s target this to release-1.9 branch.

We need to update the CHANGELOG and squash the commits.